requirements.txt
openai==1.58.1
requests==2.32.3
PyYAML==6.0.2
tqdm==4.67.1
python-dotenv==1.0.1